requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
google-generativeai>=0.5.0
aiohttp>=3.9.0
python-dateutil>=2.8.0
